Significant changes to support measurements in um
The script will now look in the MetaData directory, if it exists, for 'scope
properties XML files.  It can now read dimension and channel information from
these XML files.  This includes # of channels, what color the channel
represents, the # of dimensions, and the dimensions size in # of pixels as well
as in micrometers.

The size of the pixels is now taken into consideration and output information is
now in um^2 for area.