Removed EXCLUDE_EDGE_PARTICLES flag
This was causing some problems in brigthfield processing.  In some cases, the
whole cell ball was large enough to butt up against the ROI, which means the
brightfield particle would get excluded with this flag on.  I could enable it
for other channels, but I'm guessing it's not really eliminated many/any
spurious detections.