Fix for numerous small holes in brightfield image
This is caused by very bright spots within the cell ball.  We want to be able
to distinguish between these artificial holes and actual holes within the
cell ball.

I decided to perform two thresholding passes, a low value threshold and a high
value threshold.  Basically, we want to cut out the "gray" middle, which is
generally the plate.  Right now, the upper threshold is calculated as 95% of
the maximum intensity.  This generally works, but I think could be improved.

One idea would be to capture the upper X% of the histogram density, if the
density in the top 5% of intensities is very low.