Breakout grayscale comp, Fix binarization bug
I found a bug during the grayscale computation.  It's cumbersome to fix it
across three scripts, so I've started to break it out into a utils script.
This is easy for imagesToClouds and 3DcellStats, because they are the same.
I still need to work on it for cellStats.

The bug is that the scale bar is present while computing the threshold value.
This ends up skewing the distribution, and if the fluorescent returns are very
low, then this can make it so the only thing present is the scale bar.

I used the exclusion zone params and used it to blank out that rectangle.