No longer sort labeled cloud, debug saving
By sorting the labeled output cloud, it leads to a less desirable coloring.
That is because all of the clusters with a lot of points will be close together
and hence be hard to distinguish in a normal colormap.

Also, the supervoxel cloud is saved if the debug param is set.