Did away with imgCat/images setup
Previously, I iteratd through all of the filenames to categorize them, then
added them to a VirtualStack.  The categorization was very inefficient, as it
required iterating through all filenames for each c/z/t combo.  I repalced this
with the ability to extract c/z/t from filenames.

Also, I have forgone loading virtual stacks.  Initially, it was thought that
each c/z combo would have an image stack for t's, but it's easier to just store
each image path and open the images on the fly.  This appears to be more memory
efficient.