Fixed issues related to using detector plugin
There were a few issues I fixed in the TrackMate source to make the
script work and moving to using a stand-alone detector plugin, I can't
fix these things anymore.  The major issue is the XML writing, which
requires me to declare features.

Also, I realized finally that the ParticleAnalyzer can't be used
multi-threaded, so I set the threads to 1.